Implement sieve of Eratosthenes
To be able to use a lazy enumerator, I have chosen iteration instead of
recursion. This is a bit less efficient, of course. But the advantage is
that I can iterate to infinity. For the recursion I need to be able to
determine an upper limit, which - obviously - is impossible when
counting to infinity.