Optimize the whole thing for speed
Found a quite interesting algorithm here:
http://www.geraldbuehler.de/primzahlen/

Starting with 2 as the first prime number and 3 as the first number to
test, for each known prime number (at every current time), it adds the
respective prime number to itself until the sum becomes equal or greater
than the tested number. If the sum equals the number under test, that
number is not a prime number. If all the prime number sums end up to be
greater than the number to test, the number under test is a prime
number.

One final optimization; only sum up prime numbers that are less than or
equal to the square root of the number under test. This optimization has
already been made in the context of the sieve of Eratosthenes, the
probably best known algorithm to determine prime numbers.